CMS-881: Removed unnecesary synchronized block in location indexes update
This lock was held for long times when there are many locations to
update after bookie compaction. This can be even worse if the entry
locations cache is too small and we have to use LDB iterators to
find entry previous entry. The problem is that while this lock is
taken, the write cache cannot be flushed and that ends up in using
more than the configured quota, until we run OOM.